+++ matrix_lib/matr_resize.c
@@ -1,8 +1,8 @@
 #include "matrix.h"
 
-void    matr_resize(float ***matrix, int numRows, int numCols, int newRows, int newCols)
+void    matr_resize(double ***matrix, int numRows, int numCols, int newRows, int newCols)
 {
-    float **resized_matrix;
+    double **resized_matrix;
 
     resized_matrix = create_matrix(newRows, newCols);
     matr_copy(matrix[0], resized_matrix, numRows, numCols);
